refactor(media-search): media search repo to match production search
