+++ backend/src/handlers/mod.rs
@@ -23,9 +23,9 @@ pub mod tagging;
 pub mod taxonomy;
 pub mod version;
 pub mod queries {
-    pub mod media;
-    pub mod location;
     pub mod hall;
+    pub mod location;
+    pub mod media;
     pub mod pagination;
     pub mod production;
     pub mod sort;
