+++ backend/src/handlers/queries/media.rs
@@ -1,12 +1,10 @@
 use database::models::{entity_type::EntityType, media::MediaSearch};
-use o2o::o2o;
 use serde::Deserialize;
 use utoipa::IntoParams;
 
 use crate::handlers::queries::sort::Sort;
 
-#[derive(o2o, Deserialize, IntoParams)]
-#[owned_into(MediaSearch)]
+#[derive(Deserialize, IntoParams)]
 pub struct MediaSearchQuery {
     pub q: Option<String>,
     #[param(value_type = EntityType, inline, required = false)]
@@ -15,7 +13,18 @@ pub struct MediaSearchQuery {
     #[param(value_type = String, required = false)]
     pub role: Option<String>,
 
-    #[owned_into(~.map(Into::into))]
     #[param(value_type = Sort, inline, required = false)]
     pub sort: Option<Sort>,
+}
+
+impl From<MediaSearchQuery> for MediaSearch {
+    fn from(value: MediaSearchQuery) -> Self {
+        Self {
+            q: value.q,
+            entity_type: value.entity_type,
+            entity_id: value.entity_id,
+            role: value.role,
+            sort: value.sort.unwrap_or_default().into(),
+        }
+    }
 }
